mb2q: Handle same subject in a thread and no SOB in patch proper
The code only checks whether a patch file exists already, but it does not
handle the case when the same subject is there for two mails, which can
happen when there is a V$N+1 reply in the thread.

While patches which lack a SOB are correctly detected, the code fails later
because is assumes that there is always a SOB.

Fixup both.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>