mb2q: Handle notmuch message ordering correctly
Notmuch provides the messages in a thread in reverse order, except for the
message which is the first one in a thread. This causes the patch ordering
in the resulting patch queue to be reverse.

Insert the mails into the list after thread starting message instead of
queueing them at the end.

Reported-by: Kurt Kanzenbach <kurt@linutronix.de>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>