mb2q: Workaround broken mbox formats
mailbox format requires that lines starting with 'From ' in the mail body
are escaped. That's usually done as '>From '.

Some mailbox generators including patchwork fail to do that which confuses
the python mailbox parser as it considers a 'From ' line in the mail body
as start of the next message.

Work around it by parsing the binary file content of the mailbox manually
and by applying heuristics to distinguish between a 'From ' line in the
body and a valid unixfrom line which starts a new message.

Reported-by: Alexei Starovoitov <alexei.starovoitov@gmail.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>