mb2q: Fix wrong patch order in notmuch mode
mb2q is able to generate a quilt patch series from a notmuch query/tag. When
marking thread to create the corresponding series the patch order is wrong
sometimes.

The code uses a notmuch search query. However, it doesn't specify the sort order
which by default corresponds to None. None means "Do not apply a special sort
order (returns results in document id order)" [1]. Therefore, the order is not
necessarily correct.

Specify the sort order to make sure that the order is always correct.

[1] - https://notmuch.readthedocs.io/projects/notmuch-python/en/latest/query.html#notmuch.Query.SORT

Signed-off-by: Kurt Kanzenbach <kurt@linutronix.de>